Work in progress on shifting query complexity from repetitious code in the story controller to named scopes in the story model.
What works well is all the commonalities in ordering and the ijoining of status
and priority information now encapsulated in scope :properly_ordered.

What does not work well is the inclusion of parameters needed to select tagged
or selected stories.

We need to figure out how to pass parameters into scopes without losing the power of ActiveRecord::Relations